Overhaul built-in types
Each built-in type has been moved to its own class. Types are now
stored in Environment instead of StandardLibrary, which has been
removed. All built-in types are now lazily evaluated, which addresses
the recurring problem where the interpreter would crash on startup if it
attempted to load a type that has an attribute of another type that has
not yet been loaded. The process of evaluating a type is still O(1).
Lastly, the implementations for built-in and user-defined types are now
distinct. Attempting to update a shared attribute of a built-in type
will now raise a runtime error.